Major refactoring, read full message.
Major refactoring in progress of UTChem readers. UTChemAsciiReader is an abstract utility class. UTChemInputReader reads INPUT files for the mesh and well information. The other classes are concrete classes for reading various data files.

The main existing issue was that these were designed to work with unofficial releases of UTChem and would not work with UTChem 9.x. The refactoring is in the process of getting the readers to work with the official UTChem 9.3 release.