Commented out some code ->
Using SetUpdateExtent and SetWholeExtent from vtkDataSet has been deprecated. Also, it seems they were trying to do this inside of RequestData, which doesn't immediately make sense, since they were also defining RequestUpdateExtent.

Their motivation was that without this, the filter wouldn't act as an input to downstream filters.

This requires greater thought, but in the interest of compiling against PV 3.98, just commenting out for the moment.