Fixed NaN hanging ->
Concentration files with "NaN" ASCII literals were causing hangs, infinite loops. I'm not sure why the infinite loop was happening. It had something to do with the input file stream being read directly into a float variable. While debugging that to even see if the "NaN"s were being read (since the infinite loop seemed to be stuck on the previous line), it stopped hanging.

Now it reads into a string, checks equality with "NaN" and either string streams into a float or sets the float equal to a quiet_NaN().