Construct QString from build number.
With gcc, Qt was choking when trying to append a build number to the window title text with the format "YYYYMMDD.##" because it interpreted that as a float. Added an explict call to QString::number() to convert it.

A bigger issue is sorting out the version numbering system RVA uses.