uncommented line ->
I had a section of code commented out, with a comment that IDW was used as an identifier in completely different contexts within INPUT. According to the UTChem 9.x User Guide, it should be only re-used on a per well basis.

I'm wondering if that message dates back to before we realized there was a difference between UTChem 9.3 and UTChem 2011_7, etc.