Major refactoring
Implemented separate ReadHeader for use by RequestInformation. Implemented RequestData to read data section directly. Added a string trim helper function for the comma delimited section of the header.

One key is that the Scalars name must be set explicitly (as opposed to the vtkpython API where an unset scalar name will default to 'Scalar_'.) The error associated with this is not helpful. It will come from a view proxy and say something about the view not being able to draw the data even thought the data exists.