Specified ISATISReaderDelegate::createPoints for vtkStructured Grids.
The original design of the ISATISReaderDelegate class passed data by
using vtkPointSet, which is a parent class to things such as
vtkStructuredGrid, vtkUnstructuredGrid, etc. This worked when the Isatis
data was being interpreted as points, because then it was a simple 1-1
transfer of Isatis coordinate info arrays into a vtkPoints object.

Now that the Isatis data is being interpreted as cells, this isn't a
simple 1-1 transfer anymore because there are more VTK_POINTS than there
are Isatis coordinates. The furthest boundary planes have to be
manufactured, plus every coordinate has to be shifted by half-deltas.

Enough information was sent to make this specific for vtkStructuredGrid.