More changes for splitting ISATIS line vs grid reading
Because lines are point-based and grids are cell-based, they need to use
different indexing (off-by-one) AND they use different methods, e.g.,
GetPointData() vs GetCellData(). This causes a series of changes.

Ideally, this would be re-architected, but as a stopgap, I added a flag
that is passed through various method calls which indicates if point- or
cell-based data is being read.