Minimum integrate pass-through
Passes a dummy value through integration. This was just to get
everything working because analysis filters like vtkIntegrateAttributes
are a little different to handle. A single vertex/point is created to
attached the result to, instead of passing through the entire dataset.
In that sense, this is a terminal filter.