Fixed UTChem 9.x ALKP loading
UTChem 9.x ALKP files weren't loading. ParaView was doing that thing
where if it fails it brings up the dialog box to select the file type.

Turns out that ALKP files were calling a particular version of the
readLayerValues method that takes a string as the first parameter.
However, they were calling it with NULL, which is not a valid empty
string constructor in this instance. Simple fix of calling with
std::string(), but took quite a while to hunt down.