Partial fix of Threshold bounds for Cut with Custom src.
The GUI for Cut Between Wells with Custom Source appears to be an
altered version of one of the ones which comes with ParaView. For some
reason, the bounds of the Scalar 1 threshold are done as
DoubleRangeDomain, while the latter is an ArrayRangeDomain.

By removing explicit min/max attributes the DoubleRangeDomain is no
longer restricted, but the sliders disappear.

TODO: alter the Scalar 1 bounds to use ArrayRangeDomain, as Scalar 2
does.